+++ src/br/com/bign/listas/ListaTag.java
@@ -14,6 +14,7 @@ import br.com.bign.nottag.SegueTag;
 
 import android.os.Bundle;
 
+import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.app.ListActivity;
 import android.content.DialogInterface;
@@ -28,6 +29,7 @@ import android.widget.ListView;
 import android.widget.Toast;
 import android.widget.AdapterView.OnItemLongClickListener;
 
+@SuppressLint("NewApi")
 public class ListaTag extends ListActivity {
 
 		NottagDAO cDao;
@@ -84,10 +86,20 @@ public class ListaTag extends ListActivity {
                          new DialogInterface.OnClickListener() {
                              public void onClick(DialogInterface dialog, int which) {
                                  /// FAZ A ExCLUSAO
-                            cDao.open();
-                            cDao.delete(c);
-                            onResume();
-                        	 dialog.dismiss();
+                            
+                            Nuvem n = new Nuvem(ListaTag.this);
+                            if(n.podeDeixarTag(c.getSegueNot()))
+                            {
+                            	 
+                            	 cDao.open();
+	                             cDao.delete(c);
+	                             onResume();
+	                        	 dialog.dismiss();
+                            }else
+	                            {
+	                            	dialog.dismiss();
+	                            	Toast.makeText(ListaTag.this, "PROBLEMA AO EXLCUIR... TENTE MAIS TARDE",Toast.LENGTH_LONG).show();
+	                            }
                              }
                          });
                  b.show();
